feat: ItemFrameProtectorを任意導入に変更
plugin.ymlのdependからsoftdependへ移動し、未導入でも起動できるように
した。IFPイベントを参照するハンドラを別リスナー(IFPEvent)へ分離し、
ItemFrameProtector導入時のみ登録する。未導入時はuseIFPを無効化する。
